+++ frontend/src/App.css
@@ -2,7 +2,6 @@
   --datapass-background-grey: #ebeff3;
   --datapass-dark-grey: #3a3a3a;
   --datapass-light-grey: #e5e5e5;
-  --text-action-blue-france: #000091;
 }
 
 header {
