--- frontend/src/components/atoms/MonComptePro/style.css
@@ -1,3 +1,0 @@
-.fr-connect-group a {
-  color: var(--text-action-blue-france);
-}
